# Optimize Enum.count/2